requirements.txt
google-genai>=1.0.0
streamlit>=1.28.0
pandas>=2.0.0
